Inject "_id" methods and values when available.
If a value responds to #id, automatically set an equivalent "_id"
key/value pair. For example, if an instance of an ActiveRecord class
Widget is set on a context object of :widget, set another attribute
called :widget_id with the value of the object's ID. This helps ensure
the context's values can be sent as a message across application
boundries.